name swapping between homepage and landing
