+++ client/src/components/ui/index.ts
@@ -7,7 +7,7 @@ export * from './HoverCard';
 export * from './Input';
 export * from './InputNumber';
 export * from './Label';
-export * from './Landing';
+export * from './Homepage';
 export * from './ModelSelect';
 export * from './Prompt';
 export * from './Slider';
